Since blocks are limited to 1024 threads each, I can't have matrices larger than 44x44 in dimension, due to how I distribute threads. Now, I'm going try and let each block handle 2 and 2 rows, over indices p and q. That way, each block requires only m-1 threads (see code as to why)

+++ diagonalisation/typeDefs.h
@@ -2,7 +2,7 @@
 #ifndef TYPEDEFS_H
 #define TYPEDEFS_H
 
-typedef float  floatType;
-//~ typedef double floatType;
+//~ typedef float  floatType;
+typedef double floatType;
 
 #endif // TYPEDEFS_H
